web/flask/requirements.txt
dateparser==1.2.1
elasticsearch==8.17.2
Flask==3.1.0
flask_cors==5.0.1
matplotlib==3.8.2
protobuf==6.30.1
python-dotenv==1.0.1
spacy==3.8.2
